Isolate TimeSignature construction to within the class
