+++ src/musicxml/time.ts
@@ -1,6 +1,6 @@
 import { NamedElement, Fraction } from '@/util';
 import { TimeSignature } from './timesignature';
-import { TIME_SYMBOLS } from './enums';
+import { TIME_SYMBOLS, TimeSymbol } from './enums';
 
 /**
  * Time represents a time signature element.
@@ -15,63 +15,34 @@ export class Time {
     return this.element.attr('number').withDefault(1).int();
   }
 
-  /** Returns the time signature of the time. */
-  getTimeSignature(): TimeSignature | null {
-    const isHidden = !!this.element.first('senza-misura');
-    if (isHidden) {
-      return TimeSignature.hidden();
-    }
-
-    // The symbol overrides any other time specifications. This is done to avoid incompatible symbol and time signature
-    // specifications.
-    const symbol = this.element.attr('symbol').enum(TIME_SYMBOLS);
-    switch (symbol) {
-      case 'common':
-        return TimeSignature.common();
-      case 'cut':
-        return TimeSignature.cut();
-    }
-
-    const beats = this.element
+  /** Returns the beats of the time. */
+  getBeats(): string[] {
+    return this.element
       .all('beats')
       .map((beats) => beats.content().str())
       .filter((content): content is string => typeof content === 'string');
-    const beatTypes = this.element
+  }
+
+  /** Returns the beat types of the time. */
+  getBeatTypes(): string[] {
+    return this.element
       .all('beat-type')
       .map((beatType) => beatType.content().str())
       .filter((content): content is string => typeof content === 'string');
-
-    const timeSignatures = new Array<TimeSignature>();
-
-    const len = Math.min(beats.length, beatTypes.length);
-    for (let index = 0; index < len; index++) {
-      const beatsPerMeasure = beats[index];
-      const beatValue = beatTypes[index];
-      const timeSignature = this.parseTimeSignature(beatsPerMeasure, beatValue);
-      timeSignatures.push(timeSignature);
-    }
-
-    if (timeSignatures.length === 0) {
-      return null;
-    }
-    if (symbol === 'single-number') {
-      return TimeSignature.singleNumber(TimeSignature.combine(timeSignatures));
-    }
-    if (timeSignatures.length === 1) {
-      return timeSignatures[0];
-    }
-    return TimeSignature.combine(timeSignatures);
   }
 
-  private parseTimeSignature(beatsPerMeasure: string, beatValue: string): TimeSignature {
-    const denominator = parseInt(beatValue.trim(), 10);
-    const numerators = beatsPerMeasure.split('+').map((b) => parseInt(b.trim(), 10));
+  /** Returns whether the time signature is hidden. */
+  isHidden(): boolean {
+    return !!this.element.first('senza-misura');
+  }
 
-    if (numerators.length > 1) {
-      const fractions = numerators.map((numerator) => new Fraction(numerator, denominator));
-      return TimeSignature.complex(fractions);
-    }
+  /** Returns the symbol of the time. */
+  getSymbol(): TimeSymbol | null {
+    return this.element.attr('symbol').enum(TIME_SYMBOLS) ?? null;
+  }
 
-    return TimeSignature.of(numerators[0], denominator);
+  /** Returns the time signature of the time. */
+  getTimeSignature(): TimeSignature | null {
+    return TimeSignature.from({ time: this });
   }
 }
